+++ Validation_Service/src/leaderElection.service.js
